+++ restclient/src/main/java/esa/restclient/BodyContent.java
@@ -1,7 +1,7 @@
 package esa.restclient;
 
 public interface BodyContent<T> {
-    int type();
+    byte type();
 
     T content();
 }
